Test: Run pytest on `hello` program
I don't currently know how to make `pytest` load the plain `hello`
file (this may requires some tweaks to the module loader) so at the
moment, so that I can keep it a normal command-line script without a
`.py` extension, I simply create a link `hello.py` before running the
tests and remove it afterwards.

I did try a `pytest.ini` with

    [pytest]
    python_files=hello

But that didn't do the trick; it's still only loading things with
`.py` extensions.